Release Candidate 1
- We now have every major feature of our project working together
- MajorInfo should now be able to fetch information from the MajorInformation table
- Fixed some styling issues with divs not stretching to every corner because of a margin on the body tag
- Fixed a bug that took me an hour to figure out. It was an extra comma in the init.sql file
- Made it so the footer is below the content on a page since before it was hovering over it which blocked a good amount of the page's content (still blocks some buttons on pages that are more than 100vh long so we'll come back to it later)
