+++ client/src/assets/root.css
@@ -1,3 +1,16 @@
+#root {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+body {
+  margin: 0;
+  overflow: visible;
+  width: 100%;
+  height: 100%;
+}
+
 .nav {
   background-color: #333;
   padding: 10px 0;
@@ -43,11 +56,10 @@
 
 footer {
   background-Color: 'lightgray';
+  margin-top: auto;
   padding: 20px;
   text-Align: 'center'; 
-  position: fixed;
   bottom: 0;
   left: 0;
   width: 100%;
-  text-align: center;
 }
