+++ client/src/explore/minigames/BioengineeringGame.css
@@ -7,6 +7,7 @@
   box-sizing: border-box;
   text-align: center;
   font-size: calc(1vw + 1vh);
+  color:black;
 }
 
 .start-screen h2,
