+++ client/src/index.css
@@ -14,7 +14,7 @@
 }
 
 .nav {
-  background: #f9f9f9;  /* temporary code just used as an example for how nav-pet in pet_styles.css adopts styling from nav unless otherwise specified*/
+  background: #f9f9f9;  /* temporary code just used as an example for how nav-pet in pet_styles.css adopts styling from nav unless otherwise specified
 }
 
 a {
@@ -35,35 +35,11 @@ body {
   background-color: #1a1a1a;
   overflow: visible;
 }
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
+*/
+@media (prefers-color-scheme: dark) {
   :root {
-    color: #213547;
-    background-color: #ffffff;
+    color: #f9f9f9;
+    background-color: #242424;
   }
   a:hover {
     color: #747bff;
@@ -71,4 +47,4 @@ button:focus-visible {
   button {
     background-color: #f9f9f9;
   }
-
+}
