+++ client/src/main.jsx
@@ -5,11 +5,10 @@ import { Root, ErrorPage } from './Root';
 import LoginComponent from './authentication/LoginComponent';
 import Quiz from './quiz/Quiz';
 import RegisterComponent from './authentication/RegisterComponent';
-import Explore, { Majorinfo } from './explore/Explore';
+import Explore, { MajorInfo } from './explore/Explore';
 import { AuthProvider } from './authentication/AuthComponent';
 import Profile from './ProfileComponent';
-import Pet from './Pet'
-import MajorDetails from './explore/MajorDetails';  // Just for testing minigame loading, change to actual major details component later
+import Pet from './pet/Pet';
 import MinigameLoader from './explore/minigames/MinigameLoader';
 
 import './index.css';
@@ -35,7 +34,7 @@ const router = createBrowserRouter([
       },
       {
         path: "explore/:majorName",
-        element: <MajorDetails />,
+        element: <MajorInfo />,
       },
       {
         path: "explore/:majorName/minigame",
