+++ client/src/pet/Pet.jsx
@@ -6,9 +6,10 @@ import './pet_styles.css';
 const Pet = () => {
     const { accessToken } = useAuth();  // Get the access token from the AuthProvider to make authenticated requests
 
-    // State to manage the pet's menu
+    // State to manage the pet's menu and skills
     const [showMenu, setShowMenu] = useState(false);
     const [menuPosition, setMenuPosition] = useState({ x: 0, y: 0 });
+    const [petStats, setPetStats] = useState({ pet_name: "O'malley", mood: "neutral", love: 0, recreation: 0, hunger: 0, cleanliness: 0 })
 
     const [position, setPosition] = useState({
         left: 80,
@@ -33,6 +34,48 @@ const Pet = () => {
 
     const [targetPosition, setTargetPosition] = useState(pickRandomPoint());
 
+    // Set the style for the body page of the pet component then reset it when the component unmounts
+    useEffect(() => {
+        document.body.style.backgroundColor = 'rgb(0, 0, 0)' // Code for grass color: 'rgb(83, 172, 15)' 
+        document.body.style.overflow = 'hidden'
+
+        return () => {
+            document.body.style.backgroundColor = ''
+            document.body.style.overflow = ''
+        }
+    }, []);
+
+    // Fetch the pet's stats from the server
+    useEffect(() => {
+        const fetchPetStats = async () => {
+            if (!accessToken) {
+                console.error('User token not found');
+                return;
+            }
+
+        // Send a request to the server to get the pet's stats
+        const response = await fetch('http://localhost:8000/api/pet/stats', {
+            method: 'GET',
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${accessToken}`, // Send the access token in the header
+            },
+        });
+
+        // Check if the request was successful (response code 200-299), then get the pet's stats
+        if (response.ok) {
+            const data = await response.json();
+            console.log(data.petStats);
+            setPetStats(data.petStats);
+        } else {
+            console.error('Error status: ${response.status}');
+        }
+        };
+        fetchPetStats();
+    }, [accessToken]);
+            
+    
+    // Move the pet to the target position
     useEffect(() => {
         const movePet = () => {
             if (showMenu) setIsWalking(false); // If the menu is open, don't move the pet (wait for the menu to close)
@@ -161,10 +204,11 @@ const Pet = () => {
             body: JSON.stringify({ interactionType }),
         });
 
-        // Check if the request was successful (response code 200-299)
+        // Check if the request was successful (response code 200-299), then update the pet's stats
         if (response.ok) {
             const data = await response.json();
             console.log(data.message, data.pet);
+            setPetStats(data.pet);
         } else {
             console.error('Error status: ${response.status}');
         }
@@ -178,11 +222,19 @@ const Pet = () => {
                 onClick={handlePetClick}
                 style={{
                     position: 'absolute',
-                    left: `${position.left}vw`,
-                    top: `${position.top}vh`,
+                    left: `${position.left}%`,
+                    top: `${position.top}%`,
                     transform: `scale(${position.scale}) scaleX(${position.flip})`,
                 }}
             ></div>
+            <div className="statBoard">
+                <h3>Pet Stats</h3>
+                <p>Mood: {petStats.mood}</p>
+                <p>Love: {petStats.love}</p>
+                <p>Recreation: {petStats.recreation}</p>
+                <p>Hunger: {petStats.hunger}</p>
+                <p>Cleanliness: {petStats.cleanliness}</p>
+            </div>
             {showMenu && <PetMenu x={menuPosition.x} y={menuPosition.y} onOptionSelected={handleOptionSelected} />}
         </div>
     );
