+++ client/src/pet/pet_styles.css
@@ -1,7 +1,4 @@
-body {
-    margin: 0%;
-    overflow: hidden;
-}
+/* Note: Even though the page resizes, we'll likely still need to add media queries to make sure the pet is always visible and the menu is always in the same place. */
 
 nav.nav-pet {
   display: flex;
@@ -12,56 +9,79 @@ nav.nav-pet {
 }
 
 nav.nav-pet:hover {
-  background: black;
+  background: #242424;
 }  
   
 .backyard {
+  aspect-ratio: 16 / 9;
   background-image: url('../assets/Backyard.png');
   position: relative;
   width: 100vw;
-  height: 100vh;
-  background-size: cover;
+  height: auto;
+  background-size: contain;
   background-position: center;
-  min-width: 300px;
-  min-height: 300px; /* Could also do this for max width/height along with the media thing below */
+  background-repeat: no-repeat;
   margin: auto;
 }
 
-@media (max-width: 300px) {
-  /* Small window sizes will show this version, could add support for phones */
-  .backyard {
-    min-width: 100%; /* Adjust size for extra small screens */
-    background-image: url('../assets/Backyard.png'); /*If we have a limit on minimum size we could add a "please resize" warning image here instead*/
-    background-size: contain; /* Change to contain to ensure the entire image is visible */
-    background-repeat: no-repeat;
-  }
-}
-
 .pet {
-  width: 512px;
-  height: 512px;
+  width: 30vw;
+  height: 30vw;
   cursor: grab;
   background-image: url('../assets/pet_walking.png');
   /*background-color: rgba(255, 0, 0, 0.5); 
   border: 2px solid red;*/
   background-repeat: no-repeat;
+  background-size: cover;
 }
 
 .walk {
-  animation: walk_cycle 1s steps(14) infinite;
+  animation: walk_cycle 1s steps(13) infinite;
 }
 
 @keyframes walk_cycle {
   from {
-    background-position: 0px;
+    background-position: 0;
   }
   to {
-    background-position: -7168px;
-  } /* 14 frames * 512px per frame */
+    background-position: 100%;
+  }
 }
 
+/*
 .flip-horizontal {
-  transform: scaleX(-1); /* Flip horizontally */
+  transform: scaleX(-1); 
+}
+*/
+
+/* Pet Stat Board */
+.statBoard {
+  position: absolute;
+  padding: 10px;
+  background-color: white;
+  border: 2px solid #333;
+  color: black;
+  border-radius: 8px;
+  top: 10%; 
+  right: 10%; 
+  width: 200px; 
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  z-index: 20; 
+}
+
+.statBoard h3 {
+  margin-top: 0;
+  text-align: center;
+}
+
+.statBoard ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.statBoard li {
+  margin: 5px 0;
+  text-align: center;
 }
 
 
@@ -71,6 +91,7 @@ nav.nav-pet:hover {
   display: flex;
   flex-direction: column;
   align-items: center;
+  z-index: 10;
 }
 
 .menu-option {
@@ -94,14 +115,14 @@ nav.nav-pet:hover {
   width: 40px;
   height: 60px;
   transform: translateY(-50px);
-  background-image: url('./assets/Top_menu_option.png');
+  background-image: url('../assets/Top_menu_option.png'); /* These are all the same image just rotated but I'll add custom symbols later*/
 }
 
 .right {
   width: 60px;
   height: 40px;
   transform: translateX(50px);
-  background-image: url('./assets/Right_menu_option.png');
+  background-image: url('../assets/Right_menu_option.png');
 
 }
 
@@ -109,13 +130,13 @@ nav.nav-pet:hover {
   width: 40px;
   height: 60px;
   transform: translateY(50px);
-  background-image: url('./assets/Bottom_menu_option.png');
+  background-image: url('../assets/Bottom_menu_option.png');
 }
 
 .left {
   width: 60px;
   height: 40px;
   transform: translateX(-50px);
-  background-image: url('./assets/Left_menu_option.png');
+  background-image: url('../assets/Left_menu_option.png');
 
 }
